apps/question_service/pyproject.toml
[project]
name = "question-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "fastapi[standard]>=0.119.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]
